requirements.txt
Flask
psycopg2-binary
python-dotenv
gunicorn
google-cloud-secret-manager